23/10

+++ Backend/src/main/java/com/swp391/gr3/ev_management/dto/request/ConfirmPaymentRequest.java
@@ -9,5 +9,5 @@ public class ConfirmPaymentRequest {
     private Long invoiceId;
     private PaymentType paymentMethod;
     private String status;
-    private double amount;
+    private Long amount;
 }
